# openapi/backend.yml
openapi: 3.0.0
paths:
...
  /api/device-type/count:
    get:
      description: Gibt die Anzahl aller Device-Typen zurück
      operationId: DeviceTypeController_getCount
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountDto'
        '401':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - DeviceType
  /api/device-type:
    get:
      description: Gibt alle Device-Typen zurück
      operationId: DeviceTypeController_getAll
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - name: offset
          required: false
          in: query
          schema:
            type: number
        - name: sortCol
          required: false
          in: query
          schema:
            type: string
        - name: sortDir
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceTypeDto'
        '401':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - DeviceType
    post:
      description: Erstellt einen Device-Typen
      operationId: DeviceTypeController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceTypeCreateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTypeDto'
        '401':
          description: ''
        '404':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - DeviceType
  /api/device-type/{id}:
    get:
      description: Gibt einen Device-Typen zurück
      operationId: DeviceTypeController_getOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTypeDto'
        '401':
          description: ''
        '404':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - DeviceType
    put:
      description: Aktualisiert einen  Device-Typen
      operationId: DeviceTypeController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceTypeUpdateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTypeDto'
        '401':
          description: ''
        '404':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - DeviceType
    delete:
      description: Löscht einen Device-Typen
      operationId: DeviceTypeController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '204':
          description: ''
        '401':
          description: ''
        '404':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - DeviceType
...
  /api/device/count:
    get:
      description: Gibt die Anzahl aller Geräte zurück
      operationId: DeviceController_getCount
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountDto'
        '401':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - Device
  /api/device:
    get:
      description: Gibt alle Geräte zurück
      operationId: DeviceController_getAll
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - name: offset
          required: false
          in: query
          schema:
            type: number
        - name: typeId
          required: false
          in: query
          schema:
            nullable: true
            type: number
        - name: groupId
          required: false
          in: query
          schema:
            nullable: true
            type: number
        - name: locationId
          required: false
          in: query
          schema:
            nullable: true
            type: number
        - name: sortCol
          required: false
          in: query
          schema:
            type: string
        - name: sortDir
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceDto'
        '401':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - Device
    post:
      description: Erstellt eine Geräte
      operationId: DeviceController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDto'
        '401':
          description: ''
        '404':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - Device
  /api/device/{id}:
    get:
      description: Gibt eine Geräte zurück
      operationId: DeviceController_getOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDto'
        '401':
          description: ''
        '404':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - Device
    put:
      description: Aktualisiert eine Geräte
      operationId: DeviceController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDto'
        '401':
          description: ''
        '404':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - Device
    delete:
      description: Löscht eine Geräte
      operationId: DeviceController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '204':
          description: ''
        '401':
          description: ''
        '404':
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - Device
...
info:
  title: OFS-Manager
  description: ''
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: oauth2
      flows:
        implicit:
          scopes:
            openid: openid
            profile: profile
          authorizationUrl: >-
            https://127.0.0.1:20443/realms/openfiresource/protocol/openid-connect/auth
          refreshUrl: >-
            https://127.0.0.1:20443/realms/openfiresource/protocol/openid-connect/token
          tokenUrl: >-
            https://127.0.0.1:20443/realms/openfiresource/protocol/openid-connect/token
  schemas:
    CountDto:
      type: object
      properties:
        count:
          type: number
      required:
        - count
...
    DeviceTypeDto:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        notice:
          type: string
          nullable: true
        pricePerUnit:
          type: number
          nullable: true
        manufactor:
          type: string
          nullable: true
        dealer:
          type: string
          nullable: true
        visualInspectionAfterUsage:
          type: boolean
      required:
        - id
        - name
        - visualInspectionAfterUsage
    DeviceTypeCreateDto:
      type: object
      properties:
        name:
          type: string
        notice:
          type: string
          nullable: true
        pricePerUnit:
          type: number
          nullable: true
        manufactor:
          type: string
          nullable: true
        dealer:
          type: string
          nullable: true
        visualInspectionAfterUsage:
          type: boolean
      required:
        - name
        - visualInspectionAfterUsage
    DeviceTypeUpdateDto:
      type: object
      properties:
        name:
          type: string
        notice:
          type: string
          nullable: true
        pricePerUnit:
          type: number
          nullable: true
        manufactor:
          type: string
          nullable: true
        dealer:
          type: string
          nullable: true
        visualInspectionAfterUsage:
          type: boolean
      required:
        - name
        - visualInspectionAfterUsage
...
    LocationDto:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: number
        parentId:
          type: number
          nullable: true
          default: null
        children:
          nullable: true
          type: array
          items:
            type: string
        parent:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/LocationDto'
      required:
        - id
        - name
        - type
    DeviceDto:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
          nullable: true
        notice:
          type: string
          nullable: true
        manufactor:
          type: string
          nullable: true
        dealer:
          type: string
          nullable: true
        serial:
          type: string
          nullable: true
        serialManufactor:
          type: string
          nullable: true
        barcode1:
          type: string
          nullable: true
        barcode2:
          type: string
          nullable: true
        producedDate:
          format: date-time
          type: string
          nullable: true
        activeDate:
          format: date-time
          type: string
          nullable: true
        decomissionDateManufacture:
          format: date-time
          type: string
          nullable: true
        decomissionDate:
          format: date-time
          type: string
          nullable: true
        state:
          type: number
        typeId:
          type: number
          nullable: true
        type:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/DeviceTypeDto'
        groupId:
          type: number
          nullable: true
        group:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/DeviceTypeDto'
        locationId:
          type: number
          nullable: true
        location:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/LocationDto'
      required:
        - id
        - state
    DeviceCreateDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        notice:
          type: string
          nullable: true
        manufactor:
          type: string
          nullable: true
        dealer:
          type: string
          nullable: true
        serial:
          type: string
          nullable: true
        serialManufactor:
          type: string
          nullable: true
        barcode1:
          type: string
          nullable: true
        barcode2:
          type: string
          nullable: true
        producedDate:
          format: date-time
          type: string
          nullable: true
        activeDate:
          format: date-time
          type: string
          nullable: true
        decomissionDateManufacture:
          format: date-time
          type: string
          nullable: true
        decomissionDate:
          format: date-time
          type: string
          nullable: true
        state:
          type: number
        typeId:
          type: number
          nullable: true
        groupId:
          type: number
          nullable: true
        group:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/DeviceTypeDto'
        locationId:
          type: number
          nullable: true
        location:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/LocationDto'
      required:
        - state
    DeviceUpdateDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        notice:
          type: string
          nullable: true
        manufactor:
          type: string
          nullable: true
        dealer:
          type: string
          nullable: true
        serial:
          type: string
          nullable: true
        serialManufactor:
          type: string
          nullable: true
        barcode1:
          type: string
          nullable: true
        barcode2:
          type: string
          nullable: true
        producedDate:
          format: date-time
          type: string
          nullable: true
        activeDate:
          format: date-time
          type: string
          nullable: true
        decomissionDateManufacture:
          format: date-time
          type: string
          nullable: true
        decomissionDate:
          format: date-time
          type: string
          nullable: true
        state:
          type: number
        typeId:
          type: number
          nullable: true
        groupId:
          type: number
          nullable: true
        group:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/DeviceTypeDto'
        locationId:
          type: number
          nullable: true
        location:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/LocationDto'
      required:
        - state
...
security:
  - {}